+++ src/js/videojs.record.js
@@ -15,6 +15,7 @@ import RecordToggle from './controls/record-toggle';
 import RecordIndicator from './controls/record-indicator';
 import PictureInPictureToggle from './controls/picture-in-picture-toggle';
 
+import Event from './event';
 import pluginDefaultOptions from './defaults';
 import formatTime from './utils/format-time';
 import setSrcObject from './utils/browser-shim';
@@ -135,7 +136,7 @@ class Record extends Plugin {
         }
 
         // wait until player ui is ready
-        this.player.one('ready', this.setupUI.bind(this));
+        this.player.one(Event.READY, this.setupUI.bind(this));
     }
 
     /**
@@ -233,11 +234,11 @@ class Record extends Plugin {
 
                 // 'loadedmetadata' and 'loadstart' events reset the
                 // durationDisplay for the first time: prevent this
-                this.player.one('loadedmetadata', () => {
+                this.player.one(Event.LOADEDMETADATA, () => {
                     // display max record time
                     this.setDuration(this.maxLength);
                 });
-                this.player.one('loadstart', () => {
+                this.player.one(Event.LOADSTART, () => {
                     // display max record time
                     this.setDuration(this.maxLength);
                 });
@@ -261,7 +262,7 @@ class Record extends Plugin {
                     }
 
                     // prevent controlbar fadeout
-                    this.player.on('userinactive', (event) => {
+                    this.player.on(Event.USERINACTIVE, (event) => {
                         this.player.userActive(true);
                     });
 
@@ -275,10 +276,10 @@ class Record extends Plugin {
 
         // disable time display events that constantly try to reset the current time
         // and duration values
-        this.player.off('timeupdate');
-        this.player.off('durationchange');
-        this.player.off('loadedmetadata');
-        this.player.off('loadstart');
+        this.player.off(Event.TIMEUPDATE);
+        this.player.off(Event.DURATIONCHANGE);
+        this.player.off(Event.LOADEDMETADATA);
+        this.player.off(Event.LOADSTART);
 
         // display max record time
         this.setDuration(this.maxLength);
@@ -345,15 +346,15 @@ class Record extends Plugin {
                     video: false
                 };
                 // remove existing microphone listeners
-                this.surfer.surfer.microphone.un('deviceReady',
+                this.surfer.surfer.microphone.un(Event.DEVICEREADY,
                     this.deviceReadyCallback);
-                this.surfer.surfer.microphone.un('deviceError',
+                this.surfer.surfer.microphone.un(Event.DEVICEERROR,
                     this.deviceErrorCallback);
 
                 // setup new microphone listeners
-                this.surfer.surfer.microphone.on('deviceReady',
+                this.surfer.surfer.microphone.on(Event.DEVICEREADY,
                     this.deviceReadyCallback);
-                this.surfer.surfer.microphone.on('deviceError',
+                this.surfer.surfer.microphone.on(Event.DEVICEERROR,
                     this.deviceErrorCallback);
 
                 // disable existing playback events
@@ -374,7 +375,7 @@ class Record extends Plugin {
 
                             // store data and notify others
                             this.player.recordedData = event.inputBuffer;
-                            this.player.trigger('audioBufferUpdate');
+                            this.player.trigger(Event.AUDIOBUFFERUPDATE);
                         }
                     };
                 }
@@ -478,8 +479,8 @@ class Record extends Plugin {
         }
 
         // reset playback listeners
-        this.off(this.player, 'timeupdate', this.playbackTimeUpdate);
-        this.off(this.player, 'ended', this.playbackTimeUpdate);
+        this.off(this.player, Event.TIMEUPDATE, this.playbackTimeUpdate);
+        this.off(this.player, Event.ENDED, this.playbackTimeUpdate);
 
         // setup recording engine
         if (this.getRecordType() !== IMAGE_ONLY) {
@@ -513,7 +514,7 @@ class Record extends Plugin {
             }
 
             // listen for events
-            this.engine.on('recordComplete', this.engineStopCallback);
+            this.engine.on(Event.RECORDCOMPLETE, this.engineStopCallback);
 
             // audio settings
             this.engine.bufferSize = this.audioBufferSize;
@@ -618,13 +619,13 @@ class Record extends Plugin {
                 this.player.pipToggle.show();
 
                 // listen to and forward Picture-in-Picture events
-                this.mediaElement.removeEventListener('enterpictureinpicture',
+                this.mediaElement.removeEventListener(Event.ENTERPICTUREINPICTURE,
                     this.onEnterPiPHandler);
-                this.mediaElement.removeEventListener('leavepictureinpicture',
+                this.mediaElement.removeEventListener(Event.LEAVEPICTUREINPICTURE,
                     this.onLeavePiPHandler);
-                this.mediaElement.addEventListener('enterpictureinpicture',
+                this.mediaElement.addEventListener(Event.ENTERPICTUREINPICTURE,
                     this.onEnterPiPHandler);
-                this.mediaElement.addEventListener('leavepictureinpicture',
+                this.mediaElement.addEventListener(Event.LEAVEPICTUREINPICTURE,
                     this.onLeavePiPHandler);
             }
             // load stream
@@ -632,16 +633,16 @@ class Record extends Plugin {
 
             // stream loading is async, so we wait until it's ready to play
             // the stream
-            this.player.one('loadedmetadata', () => {
+            this.player.one(Event.LOADEDMETADATA, () => {
                 // start stream
                 this.mediaElement.play();
 
                 // forward to listeners
-                this.player.trigger('deviceReady');
+                this.player.trigger(Event.DEVICEREADY);
             });
         } else {
             // forward to listeners
-            this.player.trigger('deviceReady');
+            this.player.trigger(Event.DEVICEREADY);
         }
     }
 
@@ -659,7 +660,7 @@ class Record extends Plugin {
             this.player.deviceErrorCode = code;
 
             // forward error to player
-            this.player.trigger('deviceError');
+            this.player.trigger(Event.DEVICEERROR);
         }
     }
 
@@ -676,8 +677,8 @@ class Record extends Plugin {
             }
 
             // reset playback listeners
-            this.off(this.player, 'timeupdate', this.playbackTimeUpdate);
-            this.off(this.player, 'ended', this.playbackTimeUpdate);
+            this.off(this.player, Event.TIMEUPDATE, this.playbackTimeUpdate);
+            this.off(this.player, Event.ENDED, this.playbackTimeUpdate);
 
             // start preview
             switch (this.getRecordType()) {
@@ -730,7 +731,7 @@ class Record extends Plugin {
                     this.createSnapshot();
 
                     // notify UI
-                    this.player.trigger('startRecord');
+                    this.player.trigger(Event.STARTRECORD);
                     break;
 
                 case VIDEO_ONLY:
@@ -738,7 +739,7 @@ class Record extends Plugin {
                 case ANIMATION:
                 case SCREEN_ONLY:
                     // wait for media stream on video element to actually load
-                    this.player.one('loadedmetadata', () => {
+                    this.player.one(Event.LOADEDMETADATA, () => {
                         // start actually recording process
                         this.startRecording();
                     });
@@ -776,7 +777,7 @@ class Record extends Plugin {
         this.engine.start();
 
         // notify UI
-        this.player.trigger('startRecord');
+        this.player.trigger(Event.STARTRECORD);
     }
 
     /**
@@ -789,7 +790,7 @@ class Record extends Plugin {
 
             if (this.getRecordType() !== IMAGE_ONLY) {
                 // notify UI
-                this.player.trigger('stopRecord');
+                this.player.trigger(Event.STOPRECORD);
 
                 // stop countdown
                 this.player.clearInterval(this.countDown);
@@ -806,7 +807,7 @@ class Record extends Plugin {
             } else {
                 if (this.player.recordedData) {
                     // notify listeners that image data is (already) available
-                    this.player.trigger('finishRecord');
+                    this.player.trigger(Event.FINISHRECORD);
                 }
             }
         }
@@ -819,7 +820,7 @@ class Record extends Plugin {
         if (this.isRecording()) {
             // stop stream once recorded data is available,
             // otherwise it'll break recording
-            this.player.one('finishRecord', this.stopStream.bind(this));
+            this.player.one(Event.FINISHRECORD, this.stopStream.bind(this));
 
             // stop recording
             this.stop();
@@ -893,7 +894,7 @@ class Record extends Plugin {
         }
 
         // notify listeners that data is available
-        this.player.trigger('finishRecord');
+        this.player.trigger(Event.FINISHRECORD);
 
         // skip loading when player is destroyed after finishRecord event
         if (this.isDestroyed()) {
@@ -914,7 +915,7 @@ class Record extends Plugin {
 
                 // restore interaction with controls after waveform
                 // rendering is complete
-                this.surfer.surfer.once('ready', () => {
+                this.surfer.surfer.once(Event.READY, () => {
                     this._processing = false;
                 });
 
@@ -928,7 +929,7 @@ class Record extends Plugin {
                 // pausing the player so we can visualize the recorded data
                 // will trigger an async video.js 'pause' event that we
                 // have to wait for.
-                this.player.one('pause', () => {
+                this.player.one(Event.PAUSE, () => {
                     // video data is ready
                     this._processing = false;
 
@@ -939,9 +940,9 @@ class Record extends Plugin {
                     this.setDuration(this.streamDuration);
 
                     // update time during playback and at end
-                    this.on(this.player, 'timeupdate',
+                    this.on(this.player, Event.TIMEUPDATE,
                         this.playbackTimeUpdate);
-                    this.on(this.player, 'ended',
+                    this.on(this.player, Event.ENDED,
                         this.playbackTimeUpdate);
 
                     // unmute local audio during playback
@@ -980,10 +981,10 @@ class Record extends Plugin {
                 this.player.pause();
 
                 // show animation on play
-                this.on(this.player, 'play', this.showAnimation);
+                this.on(this.player, Event.PLAY, this.showAnimation);
 
                 // hide animation on pause
-                this.on(this.player, 'pause', this.hideAnimation);
+                this.on(this.player, Event.PAUSE, this.hideAnimation);
                 break;
         }
     }
@@ -1016,7 +1017,7 @@ class Record extends Plugin {
             this.setCurrentTime(currentTime, duration);
 
             // notify listeners
-            this.player.trigger('progressRecord');
+            this.player.trigger(Event.PROGRESSRECORD);
         }
     }
 
@@ -1170,15 +1171,15 @@ class Record extends Plugin {
      */
     dispose() {
         // disable common event listeners
-        this.player.off('ready');
-        this.player.off('userinactive');
-        this.player.off('loadedmetadata');
+        this.player.off(Event.READY);
+        this.player.off(Event.USERINACTIVE);
+        this.player.off(Event.LOADEDMETADATA);
 
         // prevent callbacks if recording is in progress
         if (this.engine) {
             this.engine.dispose();
             this.engine.destroy();
-            this.engine.off('recordComplete', this.engineStopCallback);
+            this.engine.off(Event.RECORDCOMPLETE, this.engineStopCallback);
         }
 
         // stop recording and device
@@ -1215,7 +1216,7 @@ class Record extends Plugin {
         // prevent callbacks if recording is in progress
         if (this.engine) {
             this.engine.dispose();
-            this.engine.off('recordComplete', this.engineStopCallback);
+            this.engine.off(Event.RECORDCOMPLETE, this.engineStopCallback);
         }
 
         // stop recording and device
@@ -1266,7 +1267,7 @@ class Record extends Plugin {
 
         // loadedmetadata resets the durationDisplay for the
         // first time
-        this.player.one('loadedmetadata', () => {
+        this.player.one(Event.LOADEDMETADATA, () => {
             // display max record time
             this.setDuration(this.maxLength);
         });
@@ -1416,10 +1417,10 @@ class Record extends Plugin {
      */
     startVideoPreview() {
         // disable playback events
-        this.off('timeupdate');
-        this.off('durationchange');
-        this.off('loadedmetadata');
-        this.off('play');
+        this.off(Event.TIMEUPDATE);
+        this.off(Event.DURATIONCHANGE);
+        this.off(Event.LOADEDMETADATA);
+        this.off(Event.PLAY);
 
         // mute local audio
         this.mediaElement.muted = true;
@@ -1519,7 +1520,7 @@ class Record extends Plugin {
         }
 
         // notify others
-        this.player.trigger('timestamp');
+        this.player.trigger(Event.TIMESTAMP);
 
         // automatically stop when max file size was reached
         if (maxFileSizeReached) {
@@ -1534,7 +1535,7 @@ class Record extends Plugin {
     enumerateDevices() {
         if (!navigator.mediaDevices || !navigator.mediaDevices.enumerateDevices) {
             this.player.enumerateErrorCode = 'enumerateDevices() not supported.';
-            this.player.trigger('enumerateError');
+            this.player.trigger(Event.ENUMERATEERROR);
             return;
         }
 
@@ -1546,10 +1547,10 @@ class Record extends Plugin {
             });
 
             // notify listeners
-            this.player.trigger('enumerateReady');
+            this.player.trigger(Event.ENUMERATEREADY);
         }).catch((err) => {
             this.player.enumerateErrorCode = err;
-            this.player.trigger('enumerateError');
+            this.player.trigger(Event.ENUMERATEERROR);
         });
     }
 
@@ -1566,7 +1567,7 @@ class Record extends Plugin {
                 // use wavesurfer
                 this.surfer.surfer.setSinkId(deviceId).then((result) => {
                     // notify listeners
-                    this.player.trigger('audioOutputReady');
+                    this.player.trigger(Event.AUDIOOUTPUTREADY);
                     return;
                 }).catch((err) => {
                     errorMessage = err;
@@ -1579,7 +1580,7 @@ class Record extends Plugin {
                     if (typeof element.sinkId !== 'undefined') {
                         element.setSinkId(deviceId).then((result) => {
                             // notify listeners
-                            this.player.trigger('audioOutputReady');
+                            this.player.trigger(Event.AUDIOOUTPUTREADY);
                             return;
                         }).catch((err) => {
                             errorMessage = err;
@@ -1594,7 +1595,7 @@ class Record extends Plugin {
         }
 
         // error if we get here: notify listeners
-        this.player.trigger('error', errorMessage);
+        this.player.trigger(Event.ERROR, errorMessage);
     }
 
     /**
@@ -1620,7 +1621,7 @@ class Record extends Plugin {
      * @param {object} event - Event data.
      */
     onEnterPiP(event) {
-        this.player.trigger('enterPIP', event);
+        this.player.trigger(Event.ENTERPIP, event);
     }
 
     /**
@@ -1629,7 +1630,7 @@ class Record extends Plugin {
      * @param {object} event - Event data.
      */
     onLeavePiP(event) {
-        this.player.trigger('leavePIP');
+        this.player.trigger(Event.LEAVEPIP);
     }
 }