+++ test/plugins/ts-ebml-plugin.spec.js
@@ -4,6 +4,8 @@
 
 import TestHelpers from '../test-helpers.js';
 
+import Event from '../../src/js/event.js';
+
 // registers the plugin
 import TsEBMLEngine from '../../src/js/plugins/ts-ebml-plugin.js';
 import {TSEBML} from '../../src/js/engine/convert-engine.js';
@@ -24,12 +26,12 @@ describe('plugins.ts-ebml-plugin', () => {
 
     /** @test {TsEBMLEngine} */
     it('converts', (done) => {
-        player.one('deviceReady', () => {
+        player.one(Event.DEVICEREADY, () => {
             let req = new Request(TestHelpers.TEST_WEBM);
             fetch(req).then((response) => {
                 return response.blob();
             }).then((blob) => {
-                player.one('finishConvert', () => {
+                player.one(Event.FINISHCONVERT, () => {
                     expect(player.convertedData instanceof Blob).toBeTruthy();
                     done();
                 });
@@ -37,7 +39,7 @@ describe('plugins.ts-ebml-plugin', () => {
             });
         });
 
-        player.one('ready', () => {
+        player.one(Event.READY, () => {
             // start device
             player.record().getDevice();
         });